make sure node_modules exist
node_modules may not exist in the parent module if all its dependencies
are installed one level up (e.g if it has no dependencies that are on a
version specific to it).

So we need to make sure that node_modules exist.

I also made everything use absolute paths, so that if an
error occurs it will be easier to understand what is going on.